@@ -1,5 +1,6 @@
 import sys
-from typing import Any, Callable, Dict, Iterable
+from collections.abc import Iterable
+from typing import Any, Callable
 
 import pytest
 
@@ -9,7 +10,7 @@ from tests.types import TestClientFactory
 
 WSGIResponse = Iterable[bytes]
 StartResponse = Callable[..., Any]
-Environment = Dict[str, Any]
+Environment = dict[str, Any]
 
 
 def hello_world(
