@@ -1,5 +1,5 @@
 import warnings
-from typing import Generator
+from collections.abc import Generator
 
 import pytest
 
