@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 import sys
-from typing import Any, Iterator
+from collections.abc import Iterator
+from typing import Any
 
 import anyio
 import pytest
