@@ -1,5 +1,6 @@
 import sys
-from typing import Any, MutableMapping
+from collections.abc import MutableMapping
+from typing import Any
 
 import anyio
 import pytest
