@@ -3,8 +3,9 @@ from __future__ import annotations
 import itertools
 import sys
 from asyncio import Task, current_task as asyncio_current_task
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from typing import Any, AsyncGenerator
+from typing import Any
 
 import anyio
 import anyio.lowlevel
